third_party/WebKit/LayoutTests/css3/flexbox/flexbox-width-with-overflow-auto.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox {
    border: 5px solid green;
    position: relative;
    height: 50px;
    display: inline-flex;
}

.overflow {
    border: 1px solid red;
    overflow: auto;
    min-width: 0;
    min-height: 0;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>
<div id=target class="flexbox" data-expected-width="47">
  <div class="overflow">
    <div style="height: 100px; width: 20px"></div>
  </div>
</div>
</body>
</html>
